New beta_B, mid_c, range calculations, & checks
This code made several modifications in bugs which were present in the
first version. For example, beta_b now actually has an equation which
solves for beta_b. Because the calculation of mid_c was apparently
random (with respect to the range of range_r/2 -> niche_n), I modified
the calculation of mid_c, minRange, and maxRange to account for that.
Additional documentation for the new variables will be added and the
current debugging fprintf statements will be removed, after additional
testing.